Extract `login_test_user_1` to a reusable utility function in `tests/utils.py` and update references in test files
