+++ tests/integration/routes/test_event_route.py
@@ -4,22 +4,7 @@ from app.models.event import Event
 from app.event.services import EventService
 from app.services.service_base import service_response
 from flask import url_for
-
-def login_test_user_1(client, test_user_1):
-    """
-    Logs a test user into the application using the provided client and credentials.
-
-    Parameters:
-    client : Any
-        The test client used to simulate HTTP requests to the application.
-    test_user_1 : Any
-        An object representing the first test user with an email attribute.
-
-    Returns:
-    Any
-        The HTTP response object resulting from the login POST request.
-    """
-    return client.post(url_for("auth.login"), data={"email": test_user_1.email, "password": "user1password"})
+from tests.utils import login_test_user_1
 
 def test_add_event_get_user_no_family(client, test_user_1):
     """
